+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -13,6 +13,7 @@ group: Deployment & Configs
 - Added context menu to 'Project Location' component.
 - Now `Types.BasePoint` bake as a named construction-plane.
 - Added 'Add Drafting View' component.
+- Added 'Element Face Paint' component.
 
 ### RC